requirements.txt
aiogram
aiosqlite
sqlalchemy
python-dotenv
asyncio
Pillow